refactor: merged json and state machine enums together
