+++ dotlottie-rs/src/lib.rs
@@ -3,7 +3,7 @@ mod fms;
 mod layout;
 mod lottie_renderer;
 mod markers;
-mod state_machine;
+mod state_machine_engine;
 mod thorvg;
 
 pub use dotlottie_player::*;
@@ -11,6 +11,6 @@ pub use fms::*;
 pub use layout::*;
 pub use lottie_renderer::*;
 pub use markers::*;
-pub use state_machine::events::*;
-pub use state_machine::*;
+pub use state_machine_engine::events::*;
+pub use state_machine_engine::*;
 pub use thorvg::*;
