+++ dotlottie-rs/src/state_machine_engine/errors.rs
