+++ dotlottie-rs/src/state_machine_engine/triggers/mod.rs
@@ -1,0 +1,10 @@
+use serde::Deserialize;
+
+#[derive(Deserialize, Debug)]
+#[serde(tag = "type")]
+pub enum Trigger {
+    Numeric { name: String, value: f32 },
+    String { name: String, value: String },
+    Boolean { name: String, value: bool },
+    Event { name: String },
+}
